move classes to v1
